Refactor chat model handling and move Ollama streaming logic to separate file
- Simplified `chooseChatModel` logic by removing redundant `isClaude` check.
- Removed unused `ClaudeChatModel` and `OllamaChatModel` implementations from `chat_main_handler.go`.
- Moved `OllamaChatModel` and its streaming logic to a new file `model_ollama_service.go`.
- Cleaned up unused code and improved code organization.
