+++ api/chat_main_handler.go
@@ -385,12 +385,7 @@ func regenerateAnswer(h *ChatHandler, w http.ResponseWriter, chatSessionUuid str
 func (h *ChatHandler) chooseChatModel(chat_session sqlc_queries.ChatSession, msgs []models.Message) ChatModel {
 	model := chat_session.Model
 	isTestChat := isTest(msgs)
-	isClaude := strings.HasPrefix(model, "claude")
-	isClaude3 := false
-	if strings.HasPrefix(model, "claude-3") {
-		isClaude = false
-		isClaude3 = true
-	}
+	isClaude3 := strings.HasPrefix(model, "claude-3")
 	isOllama := strings.HasPrefix(model, "ollama-")
 	isGemini := strings.HasPrefix(model, "gemini")
 
@@ -401,9 +396,7 @@ func (h *ChatHandler) chooseChatModel(chat_session sqlc_queries.ChatSession, msg
 	isCustom := strings.HasPrefix(model, "custom-")
 
 	var chatModel ChatModel
-	if isClaude {
-		chatModel = &ClaudeChatModel{h: h}
-	} else if isClaude3 {
+	if isClaude3 {
 		chatModel = &Claude3ChatModel{h: h}
 	} else if isTestChat {
 		chatModel = &TestChatModel{h: h}
@@ -618,271 +611,8 @@ type ClaudeResponse struct {
 	Exception  interface{} `json:"exception"`
 }
 
-func (h *ChatHandler) chatStreamClaude(w http.ResponseWriter, chatSession sqlc_queries.ChatSession, chat_compeletion_messages []models.Message, chatUuid string, regenerate bool) (*models.LLMAnswer, error) {
-	// set the api key
-	chatModel, err := h.service.q.ChatModelByName(context.Background(), chatSession.Model)
-	if err != nil {
-		RespondWithAPIError(w, ErrResourceNotFound("chat model: "+chatSession.Model))
-		return nil, err
-	}
 
-	// OPENAI_API_KEY
 
-	// create a new strings.Builder
-	// iterate through the messages and format them
-	// print the user's question
-	// convert assistant's response to json format
-	prompt := claude.FormatClaudePrompt(chat_compeletion_messages)
-	// create the json data
-	jsonData := map[string]interface{}{
-		"prompt":               prompt,
-		"model":                chatSession.Model,
-		"max_tokens_to_sample": chatSession.MaxTokens,
-		"temperature":          chatSession.Temperature,
-		"stop_sequences":       []string{"\n\nHuman:"},
-		"stream":               true,
-	}
-
-	// convert data to json format
-	jsonValue, _ := json.Marshal(jsonData)
-	// create the request
-	req, err := http.NewRequest("POST", chatModel.Url, bytes.NewBuffer(jsonValue))
-
-	if err != nil {
-		apiErr := ErrInternalUnexpected
-		apiErr.Detail = "Failed to create request"
-		apiErr.DebugInfo = err.Error()
-		RespondWithAPIError(w, apiErr)
-		return nil, err
-	}
-
-	// add headers to the request
-	apiKey := os.Getenv(chatModel.ApiAuthKey)
-	authHeaderName := chatModel.ApiAuthHeader
-	if authHeaderName != "" {
-		req.Header.Set(authHeaderName, apiKey)
-	}
-
-	req.Header.Set("Content-Type", "application/json")
-
-	// set the streaming flag
-	req.Header.Set("Accept", "text/event-stream")
-	req.Header.Set("Cache-Control", "no-cache")
-	req.Header.Set("Connection", "keep-alive")
-
-	// create the http client and send the request
-	client := &http.Client{
-		Timeout: 5 * time.Minute,
-	}
-	resp, err := client.Do(req)
-	if err != nil {
-		var apiErr APIError
-
-		// Check for specific HTTP client errors
-		if strings.Contains(err.Error(), "timeout") || strings.Contains(err.Error(), "deadline exceeded") {
-			apiErr = ErrExternalTimeout
-			apiErr.Detail = "The AI model service took too long to respond"
-			apiErr.DebugInfo = err.Error()
-		} else if strings.Contains(err.Error(), "connection refused") || strings.Contains(err.Error(), "no such host") {
-			apiErr = ErrExternalUnavailable
-			apiErr.Detail = "Could not connect to the AI model service"
-			apiErr.DebugInfo = err.Error()
-		} else {
-			apiErr = ErrInternalUnexpected
-			apiErr.Detail = "Failed to create request to AI service"
-			apiErr.DebugInfo = err.Error()
-		}
-
-		RespondWithAPIError(w, apiErr)
-		return nil, err
-	}
-
-	ioreader := bufio.NewReader(resp.Body)
-
-	// read the response body
-	defer resp.Body.Close()
-	// loop over the response body and print data
-
-	setSSEHeader(w)
-
-	flusher, ok := w.(http.Flusher)
-	if !ok {
-		apiErr := ErrInternalUnexpected
-		apiErr.Detail = "Streaming unsupported by the client"
-		RespondWithAPIError(w, apiErr)
-		return nil, err
-	}
-
-	var answer string
-	var answer_id string
-
-	if regenerate {
-		answer_id = chatUuid
-	}
-
-	var headerData = []byte("data: ")
-	count := 0
-	for {
-		count++
-		// prevent infinite loop
-		if count > 10000 {
-			break
-		}
-		line, err := ioreader.ReadBytes('\n')
-		if err != nil {
-			if errors.Is(err, io.EOF) {
-				fmt.Println("End of stream reached")
-				break // Exit loop if end of stream
-			}
-			return nil, err
-		}
-		if !bytes.HasPrefix(line, headerData) {
-			continue
-		}
-		line = bytes.TrimPrefix(line, headerData)
-
-		if bytes.HasPrefix(line, []byte("[DONE]")) {
-			// stream.isFinished = true
-			fmt.Println("DONE break")
-			data, _ := json.Marshal(constructChatCompletionStreamReponse(answer_id, answer))
-			fmt.Fprintf(w, "data: %v\n\n", string(data))
-			flusher.Flush()
-			break
-		}
-		if answer_id == "" {
-			answer_id = NewUUID()
-		}
-		var response ClaudeResponse
-		_ = json.Unmarshal(line, &response)
-		answer = response.Completion
-		if len(answer) < 200 || len(answer)%2 == 0 {
-			data, _ := json.Marshal(constructChatCompletionStreamReponse(answer_id, answer))
-			fmt.Fprintf(w, "data: %v\n\n", string(data))
-			flusher.Flush()
-		}
-	}
-
-	return &models.LLMAnswer{
-		Answer:   answer,
-		AnswerId: answer_id,
-	}, nil
-}
-
-func (h *ChatHandler) chatOllamStream(w http.ResponseWriter, chatSession sqlc_queries.ChatSession, chat_compeletion_messages []models.Message, chatUuid string, regenerate bool) (*models.LLMAnswer, error) {
-	// set the api key
-	chatModel, err := h.service.q.ChatModelByName(context.Background(), chatSession.Model)
-	if err != nil {
-		RespondWithAPIError(w, ErrResourceNotFound("chat model: "+chatSession.Model))
-		return nil, err
-	}
-	jsonData := map[string]interface{}{
-		"model":    strings.Replace(chatSession.Model, "ollama-", "", 1),
-		"messages": chat_compeletion_messages,
-	}
-	// convert data to json format
-	jsonValue, _ := json.Marshal(jsonData)
-	// create the request
-	req, err := http.NewRequest("POST", chatModel.Url, bytes.NewBuffer(jsonValue))
-
-	if err != nil {
-		RespondWithAPIError(w, ErrInternalUnexpected.WithDetail("Failed to make request").WithDebugInfo(err.Error()))
-		return nil, err
-	}
-
-	// add headers to the request
-	apiKey := os.Getenv(chatModel.ApiAuthKey)
-	authHeaderName := chatModel.ApiAuthHeader
-	if authHeaderName != "" {
-		req.Header.Set(authHeaderName, apiKey)
-	}
-
-	req.Header.Set("Content-Type", "application/json")
-
-	// set the streaming flag
-	req.Header.Set("Accept", "text/event-stream")
-	req.Header.Set("Cache-Control", "no-cache")
-	req.Header.Set("Connection", "keep-alive")
-
-	// create the http client and send the request
-	client := &http.Client{
-		Timeout: 5 * time.Minute,
-	}
-	resp, err := client.Do(req)
-	if err != nil {
-		RespondWithAPIError(w, ErrInternalUnexpected.WithDetail("Failed to create chat completion stream").WithDebugInfo(err.Error()))
-		return nil, err
-	}
-
-	ioreader := bufio.NewReader(resp.Body)
-
-	// read the response body
-	defer resp.Body.Close()
-	// loop over the response body and print data
-
-	setSSEHeader(w)
-
-	flusher, ok := w.(http.Flusher)
-	if !ok {
-		RespondWithAPIError(w, APIError{
-			HTTPCode: http.StatusInternalServerError,
-			Code:     "STREAM_UNSUPPORTED",
-			Message:  "Streaming unsupported by client",
-		})
-		return nil, err
-	}
-
-	var answer string
-	var answer_id string
-
-	if regenerate {
-		answer_id = chatUuid
-	}
-
-	count := 0
-	for {
-		count++
-		// prevent infinite loop
-		if count > 10000 {
-			break
-		}
-		line, err := ioreader.ReadBytes('\n')
-		if err != nil {
-			if errors.Is(err, io.EOF) {
-				fmt.Println("End of stream reached")
-				break // Exit loop if end of stream
-			}
-			return nil, err
-		}
-		var streamResp OllamaResponse
-		err = json.Unmarshal(line, &streamResp)
-		if err != nil {
-			return nil, err
-		}
-		answer += strings.ReplaceAll(streamResp.Message.Content, "<0x0A>", "\n")
-		if streamResp.Done {
-			// stream.isFinished = true
-			fmt.Println("DONE break")
-			data, _ := json.Marshal(constructChatCompletionStreamReponse(answer_id, answer))
-			fmt.Fprintf(w, "data: %v\n\n", string(data))
-			flusher.Flush()
-			break
-		}
-		if answer_id == "" {
-			answer_id = NewUUID()
-		}
-
-		if len(answer) < 200 || len(answer)%2 == 0 {
-			data, _ := json.Marshal(constructChatCompletionStreamReponse(answer_id, answer))
-			fmt.Fprintf(w, "data: %v\n\n", string(data))
-			flusher.Flush()
-		}
-	}
-
-	return &models.LLMAnswer{
-		Answer:   answer,
-		AnswerId: answer_id,
-	}, nil
-}
 
 type CustomModelResponse struct {
 	Completion string      `json:"completion"`
@@ -1126,15 +856,6 @@ func constructChatCompletionStreamReponse(answer_id string, answer string) opena
 	return resp
 }
 
-// Claude ChatModel implementation
-type ClaudeChatModel struct {
-	h *ChatHandler
-}
-
-func (m *ClaudeChatModel) Stream(w http.ResponseWriter, chatSession sqlc_queries.ChatSession, chat_compeletion_messages []models.Message, chatUuid string, regenerate bool, stream bool) (*models.LLMAnswer, error) {
-	return m.h.chatStreamClaude(w, chatSession, chat_compeletion_messages, chatUuid, regenerate)
-}
-
 // Test ChatModel implementation
 type TestChatModel struct {
 	h *ChatHandler
@@ -1144,14 +865,6 @@ func (m *TestChatModel) Stream(w http.ResponseWriter, chatSession sqlc_queries.C
 	return m.h.chatStreamTest(w, chatSession, chat_compeletion_messages, chatUuid, regenerate)
 }
 
-// Ollama ChatModel implementation
-type OllamaChatModel struct {
-	h *ChatHandler
-}
-
-func (m *OllamaChatModel) Stream(w http.ResponseWriter, chatSession sqlc_queries.ChatSession, chat_compeletion_messages []models.Message, chatUuid string, regenerate bool, stream bool) (*models.LLMAnswer, error) {
-	return m.h.chatOllamStream(w, chatSession, chat_compeletion_messages, chatUuid, regenerate)
-}
 
 // Completion ChatModel implementation
 type CompletionChatModel struct {
